Refactor to route, controllers, services
