+++ compiler-service/src/index.ts
@@ -1,6 +1,7 @@
 import express, { Request, Response } from 'express';
 
-import { compilerRouter } from './routes/compiler.routes';
+import router from './routes/index';
+
 const app = express();
 const PORT = 3005;
 
@@ -10,7 +11,7 @@ app.get('/', (req: Request, res: Response) => {
     res.send('Hello, World!');
 });
 
-app.use(compilerRouter)
+app.use(router)
 
 app.listen(PORT, () => {
     console.log(`Server is running on http://localhost:${PORT}`);
