+++ compiler-service/src/routes/index.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import { compilerRouter } from './compiler.routes';
+
+const router = Router();
+
+router.use(compilerRouter);
+
+export default router;
